+++ src/core/CMakeLists.txt
@@ -489,6 +489,10 @@ add_library(core STATIC
     hle/service/apm/apm_controller.h
     hle/service/apm/apm_interface.cpp
     hle/service/apm/apm_interface.h
+    hle/service/audio/audio_controller.cpp
+    hle/service/audio/audio_controller.h
+    hle/service/audio/audio_device.cpp
+    hle/service/audio/audio_device.h
     hle/service/audio/audio_in_manager.cpp
     hle/service/audio/audio_in_manager.h
     hle/service/audio/audio_in.cpp
@@ -497,12 +501,12 @@ add_library(core STATIC
     hle/service/audio/audio_out_manager.h
     hle/service/audio/audio_out.cpp
     hle/service/audio/audio_out.h
+    hle/service/audio/audio_renderer_manager.cpp
+    hle/service/audio/audio_renderer_manager.h
+    hle/service/audio/audio_renderer.cpp
+    hle/service/audio/audio_renderer.h
     hle/service/audio/audio.cpp
     hle/service/audio/audio.h
-    hle/service/audio/audio_controller.cpp
-    hle/service/audio/audio_controller.h
-    hle/service/audio/audren_u.cpp
-    hle/service/audio/audren_u.h
     hle/service/audio/errors.h
     hle/service/audio/final_output_recorder_manager_for_applet.cpp
     hle/service/audio/final_output_recorder_manager_for_applet.h
