+++ src/core/hle/service/audio/audio.cpp
@@ -6,7 +6,7 @@
 #include "core/hle/service/audio/audio_controller.h"
 #include "core/hle/service/audio/audio_in_manager.h"
 #include "core/hle/service/audio/audio_out_manager.h"
-#include "core/hle/service/audio/audren_u.h"
+#include "core/hle/service/audio/audio_renderer_manager.h"
 #include "core/hle/service/audio/final_output_recorder_manager.h"
 #include "core/hle/service/audio/final_output_recorder_manager_for_applet.h"
 #include "core/hle/service/audio/hwopus.h"
@@ -25,7 +25,8 @@ void LoopProcess(Core::System& system) {
         "audrec:a", std::make_shared<IFinalOutputRecorderManagerForApplet>(system));
     server_manager->RegisterNamedService("audrec:u",
                                          std::make_shared<IFinalOutputRecorderManager>(system));
-    server_manager->RegisterNamedService("audren:u", std::make_shared<AudRenU>(system));
+    server_manager->RegisterNamedService("audren:u",
+                                         std::make_shared<IAudioRendererManager>(system));
     server_manager->RegisterNamedService("hwopus", std::make_shared<HwOpus>(system));
     ServerManager::RunServer(std::move(server_manager));
 }
